reorganizing query
