+++ src/app/admin/fund/page.tsx
@@ -1,6 +1,6 @@
-import { getPendingBusinesses } from "~/server/fetchQuery";
 import BusinessTable from "~/components/business/business_table";
 import { Business } from "~/components/business/business_table";
+import { getPendingBusinesses } from "~/server/repository/business_repository";
 
 export default async function Fund() {
   const data = (await getPendingBusinesses()) as unknown as Business[];
