+++ src/app/api/save_user_to_db/route.ts
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 
-import { createUserForCurrentUser } from "~/server/createQuery";
+import { createUserForCurrentUser } from "~/server/repository/user_repository";
 
 export async function GET() {
   await createUserForCurrentUser();
