+++ src/app/startup_dashboard/page.tsx
@@ -11,11 +11,6 @@ import { Button } from "~/components/ui/button";
 import * as React from "react";
 
 import { Progress } from "~/components/ui/progress";
-import {
-  getBusinessByUserID,
-  getInvestmentByBusinessID,
-  getRequestByID,
-} from "~/server/fetchQuery";
 
 import {
   DataroomRequestWithUser,
@@ -24,6 +19,9 @@ import {
 import Link from "next/link";
 import { DataroomManager } from "~/components/dataroom/dataroom_manager";
 import { getDayUntilDeadline } from "~/utils/util";
+import { getBusinessByUserID } from "~/server/repository/business_repository";
+import { getInvestmentByBusinessID } from "~/server/repository/investment_repository";
+import { getRequestByIDWithUser } from "~/server/repository/dataroom_request_repository";
 
 export const dynamic = "force-dynamic";
 
@@ -45,7 +43,7 @@ export default async function StartupDashboard() {
 
   const businessID = business.businessID;
 
-  const dataroomRequests = await getRequestByID(businessID);
+  const dataroomRequests = await getRequestByIDWithUser(businessID);
 
   const validatedRequests: DataroomRequestWithUser[] = dataroomRequests.map(
     (request) => ({
