+++ src/server/action/approve_business.tsx
@@ -1,6 +1,6 @@
 "use server";
 
-import { acceptUserStatus } from "~/server/fetchQuery";
+import { acceptUserStatus } from "~/server/repository/business_repository";
 
 export async function approveBusinessAction(businessID: number) {
   try {
@@ -9,7 +9,10 @@ export async function approveBusinessAction(businessID: number) {
     if (statusResult.success) {
       return { status: 200, message: "User status updated successfully" };
     } else {
-      return { status: 500, message: statusResult.error || "Failed to update user status" };
+      return {
+        status: 500,
+        message: statusResult.error || "Failed to update user status",
+      };
     }
   } catch (error) {
     console.error("Server action error:", error);
